Added Context Menu and ListView
Neither are functional. I couldn't figure out for the life of me how to
integrate these two with the database to make them functional. However
if you go to design view you can see the screens and so that part of the
coding is done. The context menu may not be 100% but I won't know more
about it until we can get the long view display up and going. Currently
compiles and runs without errors.